wechat_cloud_api/requirements.txt
# WeChat Cloud API Service Dependencies
# 基于FastAPI的微信公众号云端API服务依赖 (Python 3.8+ 兼容版本)

# Web框架 (Python 3.8兼容版本)
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0

# HTTP客户端 (Python 3.8兼容版本)
aiohttp>=3.8.0,<4.0.0
httpx>=0.24.0,<1.0.0

# 数据验证和序列化 (使用Pydantic v1兼容Python 3.8)
pydantic>=1.8.0,<2.0.0

# 异步文件操作
aiofiles>=0.8.0,<24.0.0

# 加密和安全
cryptography>=3.4.0
python-multipart>=0.0.5

# 日志和监控 (兼容版本)
structlog>=21.5.0

# 环境变量管理
python-dotenv>=0.19.0

# 时间处理
python-dateutil>=2.8.0

# 生产环境服务器
gunicorn>=20.1.0

# 可选依赖（兼容版本）
redis>=4.0.0
prometheus-client>=0.14.0

# 开发和测试依赖
pytest>=7.0.0
pytest-asyncio>=0.19.0
requests>=2.28.0